+++ client_int_test.go
@@ -298,48 +298,3 @@ func within(t *testing.T, root string) func() {
 		rm(t, root)
 	}
 }
-
-func pwd(t *testing.T) string {
-	t.Helper()
-	dir, err := os.Getwd()
-	if err != nil {
-		t.Fatal(err)
-	}
-	return dir
-}
-
-func mkdir(t *testing.T, dir string) {
-	t.Helper()
-	if err := os.MkdirAll(dir, 0700); err != nil {
-		t.Fatal(err)
-	}
-}
-
-func cd(t *testing.T, dir string) {
-	t.Helper()
-	if err := os.Chdir(dir); err != nil {
-		t.Fatal(err)
-	}
-}
-
-func rm(t *testing.T, dir string) {
-	t.Helper()
-	if err := os.RemoveAll(dir); err != nil {
-		t.Fatal(err)
-	}
-}
-
-func touch(file string) {
-	_, err := os.Stat(file)
-	if os.IsNotExist(err) {
-		f, err := os.Create(file)
-		if err != nil {
-			panic(err)
-		}
-		defer f.Close()
-	}
-	t := time.Now().Local()
-	if err := os.Chtimes(file, t, t); err != nil {
-		panic(err)
-	}
-}
